Change PIN and manual 2step rollout
Changing the PIN is now only possible on tokens which have the PIN enabled. This is fixed on creation. closes #13

2step rollout is possible when adding a token manually.
The dialog that shows the phonepart (the salt) can only be closed by pressing OK and the first OTP is displayed to verify the process was successful. #4
Fixed a bug where selecting sha256/512 caused the app to crash. closes #8