Hashed PIN, BCrypt, some visual changes
The PIN is now hashed with HMACSHA512.
The Toolbar of the EnterDetailsActivity has a back button and matches the theme of the Settings.
Exchanged the Textoptions to chose from when adding a new token (QR/manual) to icons (which are very bad and need to be changed).
Added a BCrypt implemtation of Damien Miller (https://github.com/jeremyh/jBCrypt), still need to check if this is feasible for our use. #4