AboutActivity, Fix 2step saving, copy to clipboard
Added AboutActivity, idea: to add a new acknowledgement, add a txt file to the folder with the license (or URL) in it and the acknowledgement as the name. Those txt files will be load in a listview and by tapping on them the license is shown or the webpage is load (unfinished). #3
Deriving version number now from gradle build script.

Fixed 2step process to save the modified secret correctly and updating the OTP value after the computation finishes. #4

Copy OTP value to clipboard function added to actionmode.